refactored action and removal of eztz dependencies
