+++ app/components/Workspace/Accounts/AccountsModal/ActivateAccounts.js
@@ -6,7 +6,7 @@ import React, { Component } from 'react';
 import { NavLink } from 'react-router-dom';
 
 const { shell } = require('electron');
-const { storageName } = require('../../../../apis/config');
+const { storageName } = require('../../../../config/tezster.config');
 const LOCAL_STORAGE_NAME = storageName;
 
 class ActivateAccounts extends Component {
