+++ app/components/Workspace/Contracts/InvokeContract/index.js
@@ -9,7 +9,7 @@ import React, { Component } from 'react';
 import swal from 'sweetalert';
 
 const conseiljs = require('conseiljs');
-const { storageName } = require('../../../../apis/config');
+const { storageName } = require('../../../../config/tezster.config');
 const LOCAL_STORAGE_NAME = storageName;
 
 class InvokeContract extends Component {
