+++ app/config/tezster.config.js
@@ -1,4 +1,6 @@
 module.exports = {
+  TEZSTER_IMAGE: 'tezsureinc/tezster:1.0.3-beta',
+  TEZSTER_CONTAINER_NAME: 'tezster',
   storageName: 'tezsure',
   provider: 'http://localhost:18731',
   identities: [
