+++ app/styles/workspace/accounts.scss
@@ -114,3 +114,11 @@ button:disabled {
 input {
   border-radius: 0rem !important;
 }
+.mnemonics-container {
+  background: #dee2e6;
+  padding: 2%;
+}
+.mnemonics {
+  font-family: inherit;
+  color: black;
+}
