+++ package.json
@@ -188,7 +188,7 @@
     "bootstrap": "^4.4.1",
     "chalk": "^4.0.0",
     "concurrently": "^5.1.0",
-    "conseiljs": "^0.3.7",
+    "conseiljs": "^0.4.3",
     "cross-env": "^7.0.2",
     "cross-spawn": "^7.0.2",
     "css-loader": "^3.5.2",
